[sdk-playground/server] split utils into lib
