+++ sdk-playground/packages/server/src/lib/bitflags.ts
@@ -1,0 +1,2 @@
+export const hasFlag = (currentFlags: number, flag: number): boolean =>
+	(currentFlags & flag) > 0;
